+++ CHANGELOG.md
@@ -295,8 +295,8 @@
 [#1268]: https://github.com/deployphp/deployer/pull/1268
 [#1265]: https://github.com/deployphp/deployer/pull/1265
 [#1263]: https://github.com/deployphp/deployer/pull/1263
-[#1252]: https://github.com/deployphp/deployer/pull/1252
 [#1256]: https://github.com/deployphp/deployer/issues/1256
+[#1252]: https://github.com/deployphp/deployer/pull/1252
 [#1251]: https://github.com/deployphp/deployer/pull/1251
 [#1246]: https://github.com/deployphp/deployer/pull/1246
 [#1236]: https://github.com/deployphp/deployer/issues/1236